Move to utils.py
